Split patron profile tests and add live coverage
